requirements.txt
google-cloud-firestore==2.10.0
beautifulsoup4==4.11.1
pyppeteer==1.0.2
openai==0.26.5
fireo==2.0.0
firebase_admin==6.1.0